app.js renamed to main.js
